register columns and change config

+++ resources/database/migrations/2018_07_20_140624_create_ldap_hook_table.php
@@ -14,6 +14,7 @@ class CreateLdapHookTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
+            $table->string('username');
             $table->string('user_type');
         });
     }
@@ -25,8 +26,9 @@ class CreateLdapHookTable extends Migration
      */
     public function down()
     {
-        if(Schema::hasColumn('users','user_type')) {
+        if(Schema::hasColumn('users','user_type') && Schema::hasColumn('users','username') ) {
             Schema::table('users',function(Blueprint $table) {
+                $table->dropColumn('username');
                 $table->dropColumn('user_type');
                
             });
